+++ changelog.d/9112.misc
@@ -1,0 +1,1 @@
+Improve `UsernamePickerTestCase`.
